feat(persuratan): implement auto-close functionality for popover and mobile sidebar sheet
