+++ src/components/persuratan/mail-toolbar.tsx
@@ -79,6 +79,54 @@ export const MailToolbar = memo(
 							Tulis Baru
 						</TooltipButton>
 
+						<div className="flex items-center gap-2">
+							<div className="flex items-center gap-1">
+								<span className="text-[10px] text-muted-foreground whitespace-nowrap">
+									Surat Dari:
+								</span>
+								<Input
+									type="date"
+									className="h-7 w-28.75 text-[10px] px-1.5 py-0"
+									value={startDate}
+									onChange={(e) => handleDateChange("start", e.target.value)}
+								/>
+								<span className="text-[10px] text-muted-foreground">s.d</span>
+								<Input
+									type="date"
+									className="h-7 w-28.75 text-[10px] px-1.5 py-0"
+									value={endDate}
+									onChange={(e) => handleDateChange("end", e.target.value)}
+								/>
+							</div>
+
+							<div className="flex items-center gap-1">
+								<span className="text-[10px] text-muted-foreground whitespace-nowrap">
+									Kata Kunci:
+								</span>
+								<div className="flex items-center gap-1">
+									<Input
+										placeholder="Cari..."
+										className="h-7 w-32 text-[10px] px-2"
+										value={searchKeyword}
+										onChange={(e) => setSearchKeyword(e.target.value)}
+										onKeyDown={(e) => {
+											if (e.key === "Enter") {
+												handleSearch();
+											}
+										}}
+									/>
+									<Button
+										size="sm"
+										className="h-7 text-[11px] px-2"
+										onClick={() => handleClearSearch()}
+									>
+										<IconSearch className="size-3 mr-1" />
+										Cari
+									</Button>
+								</div>
+							</div>
+						</div>
+
 						{selectedMailId && (
 							<>
 								<DropdownMenu>
@@ -155,53 +203,6 @@ export const MailToolbar = memo(
 
 					<div className="h-4 w-px bg-border mx-1 hidden sm:block" />
 
-					<div className="flex items-center gap-2">
-						<div className="flex items-center gap-1">
-							<span className="text-[10px] text-muted-foreground whitespace-nowrap">
-								Surat Dari:
-							</span>
-							<Input
-								type="date"
-								className="h-7 w-28.75 text-[10px] px-1.5 py-0"
-								value={startDate}
-								onChange={(e) => handleDateChange("start", e.target.value)}
-							/>
-							<span className="text-[10px] text-muted-foreground">s.d</span>
-							<Input
-								type="date"
-								className="h-7 w-28.75 text-[10px] px-1.5 py-0"
-								value={endDate}
-								onChange={(e) => handleDateChange("end", e.target.value)}
-							/>
-						</div>
-
-						<div className="flex items-center gap-1">
-							<span className="text-[10px] text-muted-foreground whitespace-nowrap">
-								Kata Kunci:
-							</span>
-							<div className="flex items-center gap-1">
-								<Input
-									placeholder="Cari..."
-									className="h-7 w-32 text-[10px] px-2"
-									value={searchKeyword}
-									onChange={(e) => setSearchKeyword(e.target.value)}
-									onKeyDown={(e) => {
-										if (e.key === "Enter") {
-											handleSearch();
-										}
-									}}
-								/>
-								<Button
-									size="sm"
-									className="h-7 text-[11px] px-2"
-									onClick={() => handleClearSearch()}
-								>
-									<IconSearch className="size-3 mr-1" />
-									Cari
-								</Button>
-							</div>
-						</div>
-					</div>
 				</div>
 
 				<div className="flex items-center gap-1"></div>
